Update chewable worker and add pending rewards.
