+++ src/components/voyagestats.tsx
@@ -1,15 +1,12 @@
 import React, { PureComponent } from 'react';
-import { Table } from 'semantic-ui-react';
+import { Table, Grid, Header } from 'semantic-ui-react';
+import ItemDisplay from '../components/itemdisplay';
 import ChewableWorker from 'worker-loader!../workers/chewableWorker';
 
+import CONFIG from '../components/CONFIG';
+
 type VoyageStatsProps = {
-	ps: number,
-	ss: number,
-	others: number[],
-	currentAm: number,
-	startAm: number,
-	elapsedSeconds: number,
-	prof: number
+	voyageData: object
 };
 
 type VoyageStatsState = {
@@ -26,14 +23,36 @@ export class VoyageStats extends PureComponent<VoyageStatsProps, VoyageStatsStat
 	}
 
 	componentDidMount() {
+		const {voyageData} = this.props;
+		const score = agg => Math.floor(agg.core + (agg.range_min+agg.range_max)/2);
+		this.config = {
+			others: [],
+			startAm: voyageData.max_hp,
+			currentAm: voyageData.hp,
+			elapsedSeconds: voyageData.voyage_duration,
+		};
+
+		for (let agg of Object.values(voyageData.skill_aggregates)) {
+			let score = Math.floor(agg.core + (agg.range_min+agg.range_max)/2);
+			let skillOdds = 0.1;
+
+			if (agg.skill == voyageData.skills.primary_skill)
+				this.config.ps = score;
+			else if (agg.skill == voyageData.skills.secondary_skill)
+				this.config.ss = score;
+			else
+				this.config.others.push(score);
+
+			this.config.variance += ((agg.range_max-agg.range_min)/(agg.core + agg.range_max))*skillOdds;
+		}
+
 		const worker = new ChewableWorker();
 		worker.addEventListener('message', message => {
 			this.setState({ estimate: message.data });
 		});
-		
-		worker.postMessage(this.props);
-	}
 
+		worker.postMessage(this.config);
+	}
 
 	_formatTime(time: number) {
 		let hours = Math.floor(time);
@@ -41,11 +60,12 @@ export class VoyageStats extends PureComponent<VoyageStatsProps, VoyageStatsStat
 		return hours+"h " +minutes+"m";
 	}
 
-	render() {
+	_renderEstimate(topMsg: string) {
 		let { estimate } = this.state;
-		
+
 		if (!estimate)
 			return (null);
+
 		const renderEst = (label, refills) => {
 			const est = estimate['refills'][refills];
 			return (
@@ -58,14 +78,10 @@ export class VoyageStats extends PureComponent<VoyageStatsProps, VoyageStatsStat
 				</tr>
 			);
 		};
-		
-		const header = this.props.elapsedSeconds 
-			?  <h3>You have a voyage running. Heres its chances.</h3>
-			: <h3>These are this voyages chances</h3>;
-		
+
 		return (
 			<div>
-				{header}
+				<h3>{topMsg}</h3>
 				<Table>
 					{renderEst("Estimate", 0)}
 					{renderEst("1 Refill", 1)}
@@ -73,6 +89,66 @@ export class VoyageStats extends PureComponent<VoyageStatsProps, VoyageStatsStat
 				</Table>
 				<p>The 20 hour voyage needs {estimate['20hrrefills']} refills at a cost of {estimate['20hrdil']} dilithium.</p>
 				<small>Powered by Chewable</small>
+			</div>
+		);
+	}
+
+	_renderRewards() {
+		const pending = this.props.voyageData.pending_rewards != [];
+		const rewards = pending
+			? this.props.voyageData.pending_rewards
+			: this.props.voyageData.granted_rewards;
+		const topMsg = pending
+		 	? 'The expected rewards are'
+			: 'The rewards to collect are';
+		const maxRarity = (type, item_type) => type == 2 && item_type == 3 ? 4 : 5;
+		const hideRarity = type => type == 3;
+		const assetURL = file => {
+			let url = file === 'energy_icon'
+				? 'atlas/energy_icon.png'
+				: `${file.substring(1).replaceAll('/', '_')}`;
+
+			if (!url.match(/\.png$/))
+				url += '.png'
+			return `${process.env.GATSBY_ASSETS_URL}${url}`;
+		};
+
+		return (
+			<div>
+				<h4>{topMsg}</h4>
+				<Grid columns={5} centered padded>
+					{rewards.loot.map((entry, idx) => (
+						<Grid.Column key={idx}>
+								<Header
+									style={{ display: 'flex', cursor: 'zoom-in' }}
+									icon={
+										<ItemDisplay
+											src={assetURL(entry.icon.file)}
+											size={48}
+											rarity={entry.rarity}
+											maxRarity={maxRarity(entry.type, entry.item_type)}
+											hideRarity={hideRarity(entry.type)}
+										/>
+									}
+									content={entry.name}
+									subheader={`Got ${entry.quantity}`}
+								/>
+						</Grid.Column>
+					))}
+				</Grid>
+			</div>
+		);
+	}
+
+	render() {
+		const { voyageData } = this.props;
+		const voyState = voyageData.state;
+
+		return (
+			<div>
+				{voyState === 'started' && this._renderEstimate('You have a voyage started.')}
+				<br/>
+				{this._renderRewards(voyageData)};
 			</div>
 		);
 	}
